Remove getProposedCommitBlock
